Rename files to appropriate name, pull out pokedex styles, create utils folder
