+++ src/App.jsx
@@ -5,7 +5,7 @@ import { Route, Switch } from 'react-router-dom';
 
 const App = () => (
   <Switch>
-    <Route exact path="/" render={(props) => <Pokedex {...props} />} />
+    <Route exact path="/" render={() => <Pokedex />} />
     <Route
       exact
       path="/:pokemonId"
