+++ src/styles/PokedexStyles.js
@@ -1,0 +1,37 @@
+import { fade } from '@material-ui/core/styles';
+
+export const PokedexStyles = (theme) => ({
+  pokedexContainer: {
+    paddingTop: '20px',
+    paddingLeft: '50px',
+    paddingRight: '50px',
+  },
+  cardMedia: {
+    margin: 'auto',
+  },
+  cardContent: {
+    textAlign: 'center',
+  },
+  searchContainer: {
+    display: 'flex',
+    backgroundColor: fade(theme.palette.common.white, 0.15),
+    paddingLeft: '20px',
+    paddingRight: '20px',
+    marginTop: '5px',
+    marginBottom: '5px',
+  },
+  searchIcon: {
+    alignSelf: 'flex-end',
+    marginBottom: '5px',
+  },
+  searchInput: {
+    width: '200px',
+    margin: '5px',
+  },
+  genButton: {
+    marginLeft: '10px',
+    marginRight: '10px',
+    paddingTop: '10px',
+    paddingBottom: '10px',
+  },
+});
